[COMMON] Fix variable size overflow in KeyStorage
In the context of several NXPConfigKS class functions, a smaller,
16 bit unsigned integer variable (sizeToRead) is used to store a
32 bit response from the KeyStorage KS_GetKeyInt() function and this
causes storage bounds overflows. The size of the "sizeToRead"
variable was extended to 32 bits to be able to correctly store the
response from the KeyStorage component.

+++ src/platform/nxp/common/NXPConfigKS.cpp
@@ -174,7 +174,7 @@ CHIP_ERROR NXPConfig::ReadConfigValueStr(Key key, char * buf, size_t bufSize, si
 {
     CHIP_ERROR err;
     ks_error_t status;
-    uint16_t sizeToRead;
+    uint32_t sizeToRead;
 
     sizeToRead = bufSize;
     VerifyOrExit(ValidConfigKey(key), err = CHIP_DEVICE_ERROR_CONFIG_NOT_FOUND); // Verify key id.
@@ -196,7 +196,7 @@ CHIP_ERROR NXPConfig::ReadConfigValueBin(const char * keyString, uint8_t * buf, 
 {
     CHIP_ERROR err;
     ks_error_t status;
-    uint16_t sizeToRead;
+    uint32_t sizeToRead;
 
     sizeToRead = bufSize;
     VerifyOrExit(keyString != NULL, err = CHIP_DEVICE_ERROR_CONFIG_NOT_FOUND); // Verify key id.
